Build shared and module versions of LesHouches
Darwin/OSX Mach-O format has strict distinction between shared and module
libraries. Module (MH_BUNDLE) on Darwin is stricly for dlopen()'ing and
you are not allowed to link to it. The interface between CMSSW and ThePEG
is not implemented according to ThePEG design. Because of that we need
to link to internal ThePEG library, which is only indented to be loaded
by ThePEG run-time on demand. This change allows us to link to LesHouches
on Darwin.